Refactor the VM for reusability
